Fix typecasting in the migration

+++ store/migrations/019_fix_validator_tx_metadata.sql
@@ -3,11 +3,11 @@ UPDATE transactions
 SET
   metadata = json_build_object(
     'node_id',         metadata->>'validator_node_id',
-    'weight',          metadata->>'validator_weight',
+    'weight',          (metadata->>'validator_weight')::bigint,
     'start_time',      metadata->>'validator_start_time',
     'end_time',        metadata->>'validator_end_time',
-    'duration',        metadata->>'validator_duration',
-    'commission_rate', metadata->>'validator_commission_rate'
+    'duration',        (metadata->>'validator_duration')::bigint,
+    'commission_rate', (metadata->>'validator_commission_rate')::bigint
   )
 WHERE
   type = 'p_add_validator'
@@ -17,10 +17,10 @@ UPDATE transactions
 SET
   metadata = json_build_object(
     'node_id',    metadata->>'validator_node_id',
-    'weight',     metadata->>'validator_weight',
+    'weight',     (metadata->>'validator_weight')::bigint,
     'start_time', metadata->>'validator_start_time',
     'end_time',   metadata->>'validator_end_time',
-    'duration',   metadata->>'validator_duration'
+    'duration',   (metadata->>'validator_duration')::bigint
   )
 WHERE
   type = 'p_add_delegator'
